Update IncludesSkinSettings.xml
Fixed a couple issues. Moved the audio codec back above the ATMOS button. These 2 have to be next to each other because of the visibility condition (and because they belong together).

Also had to renable to footer for Spotlight and trailers button (and moved it back with the rest of the footer settings).